- added spinner and width restrictions
