+++ src/app.styles.scss
@@ -5,6 +5,7 @@
   display: flex;
   flex-direction: column;
   min-height: 100%;
+  height: 100%;
 }
 
 .main-content {
