+++ src/components/header/header.tsx
@@ -11,21 +11,23 @@ export interface HeaderProps {
 
 export const Header = ({ logo, leftSlot, rightSlot }: HeaderProps) => (
   <header className="header">
-    {leftSlot}
-    <div className="logo">
-      <Link
-        to={logo?.url ?? "/"}
-        target={logo?.newTab ? "_blank" : undefined}
-        className="logo-link"
-        aria-label="logo"
-      >
-        <img
-          alt={logo?.altText ?? ""}
-          src={logo?.image?.data?.attributes?.url}
-          className="logo-icon"
-        />
-      </Link>
+    <div className="header-content">
+      {leftSlot}
+      <div className="logo">
+        <Link
+          to={logo?.url ?? "/"}
+          target={logo?.newTab ? "_blank" : undefined}
+          className="logo-link"
+          aria-label="logo"
+        >
+          <img
+            alt={logo?.altText ?? ""}
+            src={logo?.image?.data?.attributes?.url}
+            className="logo-icon"
+          />
+        </Link>
+      </div>
+      {rightSlot}
     </div>
-    {rightSlot}
   </header>
 );
