+++ src/features/image-banner/utils.ts
