+++ src/features/query/query.scss
@@ -1,0 +1,7 @@
+.query-spinner {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
