+++ src/features/wave-banner/wave-banner.scss
